edit order by issue add button in add manufacturer supplier department add

+++ app/Controller/UsersController.php
@@ -109,18 +109,27 @@ class UsersController extends AppController {
 		} else {
 			$idusr = '0001';
 		}
-        $this->request->data['User']['id'] = $idusr;
-
-        if ($this->request->is('post')) {
-            $this->User->create();
-            if ($this->User->save($this->request->data)) {
-                $currentUser = $this->Session->read('Auth.User');
-                $this->request->data['Users']['created_by'] = $currentUser['username'];
-                $this->Session->setFlash(__('The user has been saved.'), 'default', array('class' => 'alert alert-success'));
-                return $this->redirect(array('action' => 'index'));
-            } else {
-                $this->Session->setFlash(__('The user could not be saved. Please, try again.'), 'default', array('class' => 'alert alert-danger'));
-            }
+		$this->request->data['User']['id'] = $idusr;
+        
+        
+		if ($this->request->is('post')) {
+			
+			$useradd =  $this->request->data['User']['username'];
+			$chkuseradd = $this->User->find('first',array('conditions' => array('User.username'=> trim($useradd))));
+			if(empty($chkuseradd)){
+				$this->User->create();
+				if ($this->User->save($this->request->data)) {
+					$currentUser = $this->Session->read('Auth.User');
+					$this->request->data['Users']['created_by'] = $currentUser['username'];
+					$this->Session->setFlash(__('The user has been saved.'), 'default', array('class' => 'alert alert-success'));
+					return $this->redirect(array('action' => 'index'));
+				}else {
+					$this->Session->setFlash(__('The user could not be saved. Please, try again.'), 'default', array('class' => 'alert alert-danger'));
+				}
+			}else{
+				$this->Session->setFlash(__('This Username could be used REASON:This username has been used already'), 'default', array('class' => 'alert alert-danger'));
+			}
+			
         }
 
         $userGroups = $this->User->UserGroup->find('list', array('empty' => '(Please Select)',
@@ -184,11 +193,16 @@ class UsersController extends AppController {
             throw new NotFoundException(__('Invalid user'));
         }
         $this->request->allowMethod('post', 'delete');
-        if ($this->User->delete()) {
-            $this->Session->setFlash(__('The user has been deleted.'), 'default', array('class' => 'alert alert-success'));
-        } else {
-            $this->Session->setFlash(__('The user could not be deleted. Please, try again.'), 'default', array('class' => 'alert alert-danger'));
+//		debug($id);die;
+		if($id == '0001'){
+			$this->Session->setFlash(__('The user could not be deleted. '), 'default', array('class' => 'alert alert-danger'));
+		}else{
+			if ($this->User->delete()) {
+				$this->Session->setFlash(__('The user has been deleted.'), 'default', array('class' => 'alert alert-success'));
+			} else {
+				$this->Session->setFlash(__('The user could not be deleted. Please, try again.'), 'default', array('class' => 'alert alert-danger'));
         }
+		}
         return $this->redirect(array('action' => 'index'));
     }
 
@@ -197,7 +211,9 @@ class UsersController extends AppController {
             $userIds = array();
 
             foreach ($this->request->data['User']['id'] as $ids) {
-                $userIds[] = $ids;
+				if($ids!='0001'){
+					$userIds[] = $ids;
+				}
             }
 			if($this->User->deleteAll(array('id' => $userIds),false)){
 				$this->User->deleteAll(array('id' => $userIds), false);
